[REF] jDriveClonr-2: Refactored and removed stale code
