+++ src/main/java/com/ibrasoft/jdriveclonr/export/GoogleSheetsExporter.java
@@ -6,8 +6,8 @@ import com.google.api.services.sheets.v4.model.Sheet;
 import com.google.api.services.sheets.v4.model.Spreadsheet;
 import com.google.common.util.concurrent.RateLimiter;
 import com.ibrasoft.jdriveclonr.model.DriveItem;
-import com.ibrasoft.jdriveclonr.model.ExportFormat;
-import com.ibrasoft.jdriveclonr.model.GoogleMime;
+import com.ibrasoft.jdriveclonr.model.mime.ExportFormat;
+import com.ibrasoft.jdriveclonr.model.mime.GoogleMime;
 import com.ibrasoft.jdriveclonr.utils.FileUtils;
 import lombok.AllArgsConstructor;
 import lombok.Data;
