+++ src/main/java/com/ibrasoft/jdriveclonr/export/GoogleSlidesExporter.java
@@ -7,8 +7,8 @@ import com.google.api.services.slides.v1.model.Presentation;
 import com.google.api.services.slides.v1.model.Thumbnail;
 import com.google.common.util.concurrent.RateLimiter;
 import com.ibrasoft.jdriveclonr.model.DriveItem;
-import com.ibrasoft.jdriveclonr.model.ExportFormat;
-import com.ibrasoft.jdriveclonr.model.GoogleMime;
+import com.ibrasoft.jdriveclonr.model.mime.ExportFormat;
+import com.ibrasoft.jdriveclonr.model.mime.GoogleMime;
 import com.ibrasoft.jdriveclonr.utils.FileUtils;
 import lombok.AllArgsConstructor;
 import lombok.Data;
